+++ src/Equinor.ProCoSys.IPO.Query/GetPersonsWithPrivileges/GetPersonsWithPrivilegesQueryHandler.cs
@@ -8,7 +8,7 @@ using Equinor.ProCoSys.IPO.Query.GetPersons;
 using MediatR;
 using ServiceResult;
 
-namespace Equinor.ProCoSys.IPO.Query.GetPersonsInUserGroup
+namespace Equinor.ProCoSys.IPO.Query.GetPersonsWithPrivileges
 {
     public class GetPersonsWithPrivilegesQueryHandler : IRequestHandler<GetPersonsWithPrivilegesQuery, Result<List<ProCoSysPersonDto>>>
     {
